# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API with methods for getting APOD.",
        "title": "NASA APOD API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/bydate": {
            "get": {
                "description": "Get APOD by the specified date",
                "produces": [
                    "application/json"
                ],
                "summary": "Get APOD by date",
                "operationId": "get-apod-by-date",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The date in YYYY-MM-DD format",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.APOD"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/list": {
            "get": {
                "description": "Get a list of all APODs",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all APODs",
                "operationId": "get-all-apods",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.APOD"
                            }
                        }
                    }
                }
            }
        },
        "/today": {
            "get": {
                "description": "Get APOD for the current date",
                "produces": [
                    "application/json"
                ],
                "summary": "Get APOD for today",
                "operationId": "get-apod-for-today",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.APOD"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.APOD": {
            "type": "object",
            "properties": {
                "copyright": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "explanation": {
                    "type": "string"
                },
                "media_type": {
                    "type": "string"
                },
                "service_version": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}